refactor[SP2-25]: Перевод в функциональный компонент
