+++ .eslintrc.json
@@ -27,6 +27,8 @@
         "prettier"
     ],
     "rules": {
+        "react/function-component-definition": [2, { "namedComponents": "arrow-function" }],
+        "no-shadow": "off",
         "react/no-unescaped-entities": 0,
         "react/react-in-jsx-scope": "off",
         "react/jsx-filename-extension": 2,
