+++ src/components/App.jsx
@@ -8,7 +8,7 @@ import { APP_LOAD, REDIRECT } from '../constants/actionTypes';
 import Article from './Article';
 import Editor from './Editor';
 import Home from './Home';
-import Login from './Login';
+import Login from './Auth/Login/Login';
 import ConnectedProfile from './Profile';
 import ProfileFavorites from './ProfileFavorites';
 import Register from './Register';
